refactor: restructure services with domain-driven architecture
Major service layer improvements:
- Replace abstract 'modules/' with intuitive domain structure
- Create 4 clear domains: llm/, document/, infrastructure/, testing/
- Move 8 service files to appropriate domain directories
- Create domain-specific index.ts files for clean imports
- Update all import statements across the application
- Remove empty mock.service.ts file
- Maintain backward compatibility with legacy imports

New structure enables intuitive imports:
- import { streamLLMQuery, queryLLM } from '../api/llm'
- import { getStreamingConfig } from '../api/infrastructure'
- import { getMockDocument } from '../api/testing'

Benefits:
- Clear domain boundaries and responsibilities
- Easier navigation and code discovery
- Simplified import statements
- Better scalability for new features
- Self-documenting service organization
