+++ src/services/api/llm/index.ts
@@ -1,0 +1,40 @@
+/**
+ * LLM Domain Services
+ * 
+ * LLM（Large Language Model）関連のサービスを統合
+ */
+
+// LLMコア機能
+export {
+  sendLLMQuery,
+  getLLMCapabilities,
+  getLLMTemplates,
+  formatPrompt
+} from './core.service'
+
+// 会話履歴管理
+export {
+  createConversationWithSystemPrompt,
+  addUserMessageToConversation,
+  createConversationWithUserMessage,
+  filterConversationHistory,
+  createConversationSummary
+} from './conversation.service'
+
+// ストリーミング機能
+export {
+  streamLLMQuery,
+  createStreamingStatsTracker,
+  type StreamingCallbacks,
+  type StreamingStats
+} from './streaming.service'
+
+// MCPツール関連
+export {
+  getMCPTools,
+  getMCPToolInfo,
+  shouldUseMCPTools,
+  integrateMCPToolResults,
+  sendLLMQueryWithTools,
+  streamLLMQueryWithTools
+} from './tools.service'
