+++ src/services/api/testing/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Testing Domain Services
+ * 
+ * テスト関連のサービスを統合
+ */
+
+// モック機能
+export {
+  getMockDocument,
+  getMockLLMResponseWithContext,
+  getMockLLMResponse,
+  getMockLLMCapabilities,
+  getMockLLMTemplates,
+  getMockFormattedPrompt
+} from './mock.service'
